Don't use versioned prefix in XSD schema url.
From https://jira.springsource.org/browse/SPR-9427:

"We generally recommend using the "schemaless" XSDs, e.g. spring-beans.xsd
instead of spring-beans-3.0.xsd. These are always mapped to the current version
of the schema within your Spring jars, so it's going to be correct across
upgrades, etc. The only time you should need to refer to an explicitly
versioned schema is for example when you want to refer to an attribute that has
been deprecated (i.e. removed) in a newer version of the same schema, which is
quite rare anyway."

No functional changes.