Fix error on bad credentials.
Fix the following exception:
java.lang.IllegalStateException: No thread-bound request found: Are you referring to request
attributes outside of an actual web request, or processing a request outside of the originally
receiving thread? If you are actually operating within a web request and still receive this message,
your code is probably running outside of DispatcherServlet/DispatcherPortlet: In this
case, use RequestContextListener or RequestContextFilter to expose the current request.