Fix 404 when serving CatalogUtils.js and application running as mvn spring-boot:run.
Before we serve JS files from /WEB-INF/classes and it works when we are running application as JAR,
because this JAR contains /WEB-INF/classes directory. But it doesn't work because when we run
application with mvn spring-boot:run command there is no WEB-INF directory inside classpath (that is
equals to target/classes). I changed /WEB-INF/classes directory to root of classpath that should
work for a both cases.

Fix #126