Enable resource versioning for CSS/JS files.
See also: http://www.mscharhag.com/spring/resource-versioning-with-spring-mvc

While it's also possible to configure resource versioning by using
spring.resources.chain.strategy.fixed.* properties, I decided to
configure it in the Url class. The first approach requires to modify
application-prod.properties file each time when resources got changed.
This means that in our setup we will have to modify this file on the
server on each deploy because we have application-prod.properties file
that resides in the separate directory on the filesystem.

Fix #235