src/test/robotframework: speed up tests by not opening pages before each test.
Open page that will be tested once in the beginning. This means that
subsequent tests will see a state from the previous tests but in
other hand it reduces number of times when we open page from 11 to 4 and
saves about 15 seconds on my laptop.

Related to #530

No functional changes.