.travis.yml: don't do deploy on pull requests to the prod branch.
Quote from http://eng.localytics.com/best-practices-and-common-mistakes-with-travis-ci/
"The most easily abused is $TRAVIS_BRANCH, particularly if you’re
determining whether or not to deploy. Be cautious though - when Travis
is building a pull request, $TRAVIS_BRANCH will be set to the target as
opposed to the origin."

No functional changes.

[ci skip]