Revert ".travis.yml: fail build when code in before_script or script blocks fails."
This reverts commit 0b3ec75e9f16f31127d444da731794afe511da95.

We shouldn't use exit in .travis.yml as it exit immediately from a script
and don't allow to the Travis CI code to be executed at all. For
example, after_failure block won't be executed in this case.

Because errexit works as exit 1 we shouldn't use it too.

Addressed to #631