Test category/country name in Russian.
This also improves an error message for the bug #633

Note that changing Россия to Чехия is very important because it also makes a bug dissappear. It
happens because these words differs in their lengths. Prior this change, name in Russian was
persisted incorrectly and represented as 6 question marks. At the same time we have Италия in the
database that also was represented as 6 question marks. That't why we had "country already exists"
error. But now we won't have it because Чехия will be represented as 5 question marks and country
will be created successfully. But with this changes, test will fail later with more readable error
message.

Addressed to #633