/series/import/request: encode URL to percent encoding prior its persisting.
The quote from URL javadoc:
%
Note, the URI class does perform escaping of its component fields in certain circumstances. The
recommended way to manage the encoding and decoding of URLs is to use URI, and to convert between
these two classes using toURI() and URI.toURL().

The URLEncoder and URLDecoder classes can also be used, but only for HTML form encoding, which is
not the same as the encoding scheme defined in RFC2396.
%

Fix #772