perf: don't open a main page just to open something, open a page that we will need instead.
Prior this change, we were opening a main page with "Open Browser" keyword and we were navigating to
a required page after that. Most of the time, the next page was a login page. Now we opening desired
page from the beginning. This reduces number of extra actions and speed ups the integration tests.
On my Mac they are now being executed on 5 seconds faster.

No functional changes.