perf: load bootstrap-languages resources from jsdelivr.net to avoid redirection.
Previously, loading of languages.png file from cdn.rawgit.com lead to a redirect to
raw.githubusercontent.com and that increased a page load time.

Anyway, we have to switch away from rawgit.com because it "is now in a sunset phase and will soon
shut down".

No functional changes.