fix: correct links to the images on prod.
A domain for static resources has been prepended to a keys instead of the values. Because of this,
when we were looking for a URL by its name, we couldn't find the value because the name in the map
was different.

Regression after dfed1c3044f38fd4f249a4801b4e7030d5696849 commit.

Fix #1055