ci: disable comments from CodeCov in pull requests.
These comments have started to look like a spam: there always 2 comments because Travis CI runs the
build twice for every PR (one for an original branch and another one when the branch is merged into
master). Also the most significant reason is that I ignore such comments because we don't have any
required level for code coverage and often we extract addition of the unit tests to a separate
issue.

Related links:
- https://docs.codecov.io/docs/pull-request-comments#section-disable-comment
- https://docs.codecov.io/docs/codecov-yaml#section-can-i-name-the-file-codecov-yml-

[skip ci]