ci(deploy.yml): temporary ignore errors from uptimerobot-related tasks.
This is another workaround for #1089 but this time we execute code and just ignore possible errors.
This allows us to check whether the issue is still reproduceable because we've already updated
Ansible for a couple of versions and it also possible that it was something related to uptimerobot
site. If the error won't happend for a couple of deploys, I'll treat it as fixed and revert this
commit.

Relate to #1089