feat: admin can add a comment to a series.
Implementation details: a comment can be added only if there is no comment yet. This is a protection
against possible race or other unforeseeable circumstances that might lead to data loss. A full
implementation of the editing can be added later so this can be considered as the first iteration.

Fix #785