feat: admin can add a release year to a series.
Implementation details: a release year can be added only if it isn't set yet. This is a protection
against possible race or other unforeseeable circumstances that might lead to data loss. A full
implementation of the editing can be added later so this can be considered as the first iteration.

Fix #1343