perf(SitemapController): replace println() by print() with an explicit newline symbol.
It turned out that println(String) invokes print() and println() in a synchronized block. Also,
println() output a line separator that depends on runtime environment. So, for example, if we run the
app under Windows, we end up with \r\n instead of \n.

Also, when we invoke println() we acquire a lock just to output a newline character. Let's write it
as part of the content instead.

A little related to #1438