chore: reduce a number of backups stored on a remote server
Prior this change we keep 3 full backups and we made a full backup every month, so efectivelly we
had backups for last 3 months. Now we make a full backup every week and keep last 4 full backups
that is 1 month.

This change fixes the issue where we exceeded 5 Gb limit on NextCloud. Unfortunately, until #424 is
not fixed, this might happen again.

[skip ci]