chore: modify tests to use different series and avoid ordering issue on PostgreSQL
Addition of new tests in 52df7f3d97689bc9449bf2c2c67d574bede570e0, reveal a difference in the order
of sales in the series list: on PostgreSQL sales without dates (NULL) are shown first.

This change apply a workaround for that issue by using different series to fix the build.

Address to #1492