test: add and configure Jest for running tests of React components
The following dependencies have been installed:
- jest: the testing framework
- babel-jest: integration for Babel (https://jestjs.io/docs/en/getting-started#using-babel),
  otherwise tests fail with "Support for the experimental syntax 'jsx' isn't currently enabled"
- @testing-library/react: the library for testing React components
- @testing-library/jest-dom: useful matches for inspecting DOM
- react and react-dom: in order to import React in tests

All the dependencies were installed by the command:
./mvnw -Pfrontend frontend:npm -Dfrontend.npm.arguments='install --save-dev jest babel-jest @testing-library/react @testing-library/jest-dom react@16.8.6 react-dom@16.8.6'

In order to execute (or execute and watch) the tests, run:
    ./mvnw -Pfrontend frontend:npm -Dfrontend.npm.arguments='test'
or
    ./mvnw -Pfrontend frontend:npm -Dfrontend.npm.arguments='run watch-test'

Alternatively, it's possible to use npx:
   cd src/main/frontend
   npx jest
or
   npx jest --watch

Note that at the moment of this commit, there is no tests yet, so the previous command will fail.

While I marked React as an external (see webpack.config.js), all bundles got additional 5 Kb and I
couldn't figure out why. Perhaps, it's how it should work, but I'm not sure as I expected that
bundles size will stay the same.

Useful links:
- https://jestjs.io
- https://github.com/testing-library/react-testing-library
- https://github.com/testing-library/jest-dom
- https://www.newline.co/@dmitryrogozhny/how-to-start-using-react-testing-library--0e7695e8

Part of #1484