feat: users can add a comment to a series
Note also that the possibility to add a comment during a series creation has been removed. A series
now can have more than one comment from different users and in order to not confuse users by the
fact that a comment is visible only to them while other fields are visible all, this possibility was
removed.

Fix #1505