ci: pin robotframework version to fix the build on TravisCI
When we don't specify robotframework version, robotframework-lint pulls the latest. This makes the
build process less stable as dependencies versions might not work in our environment or conflict
with robotframework-lint itself.

This is exactly what has happened this time: robotframework 4.0.3 pulls ruamel 0.2.4 that requires
Python 3.5 that we don't have on Travis CI (but will be available after #1467).

Fix #1536