ci: allow to run the application only to validate Liquibase migrations
Now the application can be run as
    java -jar target/mystamps.war liquibase validate
or
    ./mvnw spring-boot:run -Dspring-boot.run.arguments='liquibase,validate'
to validate that all migrations and their checksums are correct. This should prevent the case when
the application is failing to start after deploy as now we have a possibility to check that
migrations are valid prior the application is run.

Part of #383