chore: correct wrongly applied (due to Liquibase bug) to the fields (instead of the tables) comments
In the commit 89ba68c0e11d59da55404614072f9eb757484243 (#1326) I used <setColumnRemarks> in order to
put comments to the fields. It has turned out that on MySQL, Liquibase put comments on the tables
instead of the fields (see https://github.com/liquibase/liquibase/pull/942). After update of
Liquibase (#1526), it became possible to fix the bug on our data.

Fix #1408