chore: fix a warning from Ansible about Python interpeter discovery by explicitly specifying it
The warning was:
[DEPRECATION WARNING]: Distribution Ubuntu 16.04 on host my-stamps.ru should use /usr/bin/python3,
but is using /usr/bin/python for backward compatibility with prior Ansible releases. A future Ansible
release will default to using the discovered platform python for this host. See
https://docs.ansible.com/ansible/2.9/reference_appendices/interpreter_discovery.html for more information.
This feature will be removed in version 2.12. Deprecation warnings can be disabled by setting
deprecation_warnings=False in ansible.cfg.

Fix #1192