chore: increase a number of Jetty threads to fix the integration tests
Correction for 08a07ceaad78017e5c45ee7362dd3300f97a5ed5 commit.

Previously, a number of threads was enough only for handling a single client connection. When we
tried to create a series with an image that should be downloaded from the same server, we couldn't
do that as an only worker was busy -- it was waiting when a server downloads an image from itself.

Fix #1633